+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * **Breaking change**: Refactored [HexaPDF::Layout::Frame] and associated data
   structures so that the complete result of fitting a box is returned
 
+### Fixed
+
+* [HexaPDF::Layout::TextLayouter] to freeze the new items when a text fragment
+  needs to be split
+
 
 ## 0.23.0 - 2022-05-26